Remove Neighbors and use Union<Distance,int> for Precision
Relates: #4091

This commit removes the Neighbors property from Completion field data type mapping and updates
the Precision type to a Union<Distance,int> to match the context suggest query inputs.
