+++ src/Nest/Mapping/Types/Specialized/Completion/CategorySuggestContext.cs
