+++ src/Nest/Mapping/Types/Specialized/Completion/ISuggestContext.cs
