+++ src/Nest/Mapping/Types/Specialized/Completion/SuggestContextsDescriptor.cs
