+++ src/Tests/Tests.Core/ManagedElasticsearch/NodeSeeders/DefaultSeeder.cs
@@ -347,6 +347,10 @@ namespace Tests.Core.ManagedElasticsearch.NodeSeeders
 						.Category(c => c
 							.Name("color")
 						)
+						.GeoLocation(c => c
+							.Name("geo")
+							.Precision(1)
+						)
 					)
 				)
 				.Scalar(p => p.NumberOfCommits, n => n.Store())
